index_podcast/requirements.txt
Django==2.2
django-allauth==0.39.1
psycopg2-binary
gunicorn